Update BackupApp UserControls
