+++ RoboSharp.BackupApp/UserControls/CopyOptionsExpander.xaml
@@ -1,0 +1,109 @@
+﻿<Expander x:Class="RoboSharp.BackupApp.UserControls.CopyOptionsExpander"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:RoboSharp.BackupApp.UserControls"
+             mc:Ignorable="d" 
+          Background="LightBlue" Header="Copy Options"
+             d:DesignHeight="450" d:DesignWidth="800">
+    <Grid Margin="10">
+        <Grid.RowDefinitions>
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+            <RowDefinition />
+        </Grid.RowDefinitions>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="Auto" />
+            <ColumnDefinition Width="5" />
+            <ColumnDefinition />
+        </Grid.ColumnDefinitions>
+
+        <TextBlock Grid.Row="0" Grid.Column="0" Text="File Filter" />
+        <TextBlock Grid.Row="1" Grid.Column="0" Text="Copy Subdirectories (/S)" />
+        <TextBlock Grid.Row="2" Grid.Column="0" Text="Copy Subdirectories Including Empty (/E)" />
+        <TextBlock Grid.Row="3" Grid.Column="0" Text="Depth (/LEV:N)" />
+        <TextBlock Grid.Row="4" Grid.Column="0" Text="Enable Restart Mode (/Z)" />
+        <TextBlock Grid.Row="5" Grid.Column="0" Text="Enable Backup Mode (/B)" />
+        <TextBlock Grid.Row="6" Grid.Column="0" Text="Enable Restart Mode With Backup Fallback (/ZB)" />
+        <TextBlock Grid.Row="7" Grid.Column="0" Text="Use Unbuffered I/O (/J)" />
+        <TextBlock Grid.Row="8" Grid.Column="0" Text="Enable EFS Raw Mode (/EFSRAW)" />
+        <TextBlock Grid.Row="9" Grid.Column="0" Text="Copy Flags (/COPY:copyflags)" />
+        <TextBlock Grid.Row="10" Grid.Column="0" Text="Copy Files With Security (/SEC)" />
+        <TextBlock Grid.Row="11" Grid.Column="0" Text="Copy All (/COPYALL)" />
+        <TextBlock Grid.Row="12" Grid.Column="0" Text="Remove File Information (/NOCOPY)" />
+        <TextBlock Grid.Row="13" Grid.Column="0" Text="Fix File Security On All Files (/SECFIX)" />
+        <TextBlock Grid.Row="14" Grid.Column="0" Text="Fix File Times On All Files (/TIMFIX)" />
+        <TextBlock Grid.Row="15" Grid.Column="0" Text="Purge (/PURGE)" />
+        <TextBlock Grid.Row="16" Grid.Column="0" Text="Mirror (/MIR)" />
+        <TextBlock Grid.Row="17" Grid.Column="0" Text="Move Files (/MOV)" />
+        <TextBlock Grid.Row="18" Grid.Column="0" Text="Move Files and Directories (/MOVE)" />
+        <TextBlock Grid.Row="19" Grid.Column="0" Text="Add Attributes (/A+:attributes)" />
+        <TextBlock Grid.Row="20" Grid.Column="0" Text="Remove Attributes (/A-:attributes)" />
+        <TextBlock Grid.Row="21" Grid.Column="0" Text="Create Directory and File Tree (/CREATE)" />
+        <TextBlock Grid.Row="22" Grid.Column="0" Text="Fat Files (/FAT)" />
+        <TextBlock Grid.Row="23" Grid.Column="0" Text="Turn Long Path Support Off (/256)" />
+        <TextBlock Grid.Row="24" Grid.Column="0" Text="Monitor Source Changes Limit (/MON:N)" />
+        <TextBlock Grid.Row="25" Grid.Column="0" Text="Monitor Source Time Limit (/MOT:N)" />
+        <TextBlock Grid.Row="26" Grid.Column="0" Text="RunHours (/RH:hhmm-hhmm)" />
+
+        <GridSplitter Width="5" Grid.Row="0" Grid.RowSpan="100" Grid.Column="1" ResizeBehavior="PreviousAndNext" ResizeDirection="Columns" />
+
+        <TextBox Grid.Row="0" Grid.Column="2" x:Name="FileFilter" Margin="10,0,0,0" />
+        <CheckBox Grid.Row="1" Grid.Column="2" x:Name="CopySubDirectories" Margin="10,0,0,0" />
+        <CheckBox Grid.Row="2" Grid.Column="2" x:Name="CopySubdirectoriesIncludingEmpty" Margin="10,0,0,0" />
+        <TextBox Grid.Row="3" Grid.Column="2" x:Name="Depth" Margin="10,0,0,0" PreviewTextInput="IsNumeric_PreviewTextInput" Width="50" HorizontalAlignment="Left" />
+        <CheckBox Grid.Row="4" Grid.Column="2" x:Name="EnableRestartMode" Margin="10,0,0,0" />
+        <CheckBox Grid.Row="5" Grid.Column="2" x:Name="EnableBackupMode" Margin="10,0,0,0" />
+        <CheckBox Grid.Row="6" Grid.Column="2" x:Name="EnableRestartModeWithBackupFallback" Margin="10,0,0,0" />
+        <CheckBox Grid.Row="7" Grid.Column="2" x:Name="UseUnbufferedIo" Margin="10,0,0,0" />
+        <CheckBox Grid.Row="8" Grid.Column="2" x:Name="EnableEfsRawMode" Margin="10,0,0,0" />
+        <TextBox Grid.Row="9" Grid.Column="2" x:Name="CopyFlags" Margin="10,0,0,0" PreviewTextInput="IsCopyFlag_PreviewTextInput" Width="100" HorizontalAlignment="Left" />
+        <CheckBox Grid.Row="10" Grid.Column="2" x:Name="CopyFilesWithSecurity" Margin="10,0,0,0" />
+        <CheckBox Grid.Row="11" Grid.Column="2" x:Name="CopyAll" Margin="10,0,0,0" />
+        <CheckBox Grid.Row="12" Grid.Column="2" x:Name="RemoveFileInformation" Margin="10,0,0,0" />
+        <CheckBox Grid.Row="13" Grid.Column="2" x:Name="FixFileSecurityOnAllFiles" Margin="10,0,0,0" />
+        <CheckBox Grid.Row="14" Grid.Column="2" x:Name="FixFileTimesOnAllFiles" Margin="10,0,0,0" />
+        <CheckBox Grid.Row="15" Grid.Column="2" x:Name="Purge" Margin="10,0,0,0" />
+        <CheckBox Grid.Row="16" Grid.Column="2" x:Name="Mirror" Margin="10,0,0,0" />
+        <CheckBox Grid.Row="17" Grid.Column="2" x:Name="MoveFiles" Margin="10,0,0,0" />
+        <CheckBox Grid.Row="18" Grid.Column="2" x:Name="MoveFilesAndDirectories" Margin="10,0,0,0" />
+        <TextBox Grid.Row="19" Grid.Column="2" x:Name="AddAttributes" Margin="10,0,0,0" PreviewTextInput="IsAttribute_PreviewTextInput" Width="100" HorizontalAlignment="Left" />
+        <TextBox Grid.Row="20" Grid.Column="2" x:Name="RemoveAttributes" Margin="10,0,0,0" PreviewTextInput="IsAttribute_PreviewTextInput" Width="100" HorizontalAlignment="Left" />
+        <CheckBox Grid.Row="21" Grid.Column="2" x:Name="CreateDirectoryAndFileTree" Margin="10,0,0,0" />
+        <CheckBox Grid.Row="22" Grid.Column="2" x:Name="FatFiles" Margin="10,0,0,0" />
+        <CheckBox Grid.Row="23" Grid.Column="2" x:Name="TurnLongPathSupportOff" Margin="10,0,0,0" />
+        <TextBox Grid.Row="24" Grid.Column="2" x:Name="MonitorSourceChangesLimit" Margin="10,0,0,0" PreviewTextInput="IsNumeric_PreviewTextInput" Width="50" HorizontalAlignment="Left" />
+        <TextBox Grid.Row="25" Grid.Column="2" x:Name="MonitorSourceTimeLimit" Margin="10,0,0,0" PreviewTextInput="IsNumeric_PreviewTextInput" Width="50" HorizontalAlignment="Left" />
+        <TextBox Grid.Row="26" Grid.Column="2" x:Name="RunHoursStartTime" Margin="10,0,0,0" PreviewTextInput="IsNumeric_PreviewTextInput" Width="50" HorizontalAlignment="Left" />
+        <TextBlock Grid.Row="26" Grid.Column="2" Text="-" Margin="62,2,0,0" />
+        <TextBox Grid.Row="26" Grid.Column="2" x:Name="RunHoursEndTime" Margin="70,0,0,0" PreviewTextInput="IsNumeric_PreviewTextInput" Width="50" HorizontalAlignment="Left" />
+    </Grid>
+</Expander>
