+++ RoboSharp.BackupApp/UserControls/JobHistoryExpander_SingleJob.xaml.cs
