+++ RoboSharp.BackupApp/UserControls/LoggingOptionsExpander.xaml
@@ -1,0 +1,37 @@
+﻿<Expander x:Class="RoboSharp.BackupApp.UserControls.LoggingOptionsExpander"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:RoboSharp.BackupApp.UserControls"
+             mc:Ignorable="d" 
+             Background="LightGray" Header="Logging Options"
+             d:DesignHeight="450" d:DesignWidth="800">
+    
+        <Grid Margin="10">
+            <Grid.RowDefinitions>
+                <RowDefinition />
+                <RowDefinition />
+                <RowDefinition />
+                <RowDefinition />
+            </Grid.RowDefinitions>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="Auto" />
+                <ColumnDefinition Width="5" />
+                <ColumnDefinition />
+            </Grid.ColumnDefinitions>
+
+            <TextBlock Grid.Row="0" Grid.Column="0" Text="Verbose Output (/V)" />
+            <TextBlock Grid.Row="1" Grid.Column="0" Text="No File Size (/NS)" />
+            <TextBlock Grid.Row="2" Grid.Column="0" Text="No Progress (/NP)" />
+            <TextBlock Grid.Row="3" Grid.Column="0" Text="List Only (/L)" />
+
+            <GridSplitter Width="5" Grid.Row="0" Grid.RowSpan="100" Grid.Column="1" ResizeBehavior="PreviousAndNext" ResizeDirection="Columns" />
+
+            <CheckBox Grid.Row="0" Grid.Column="2" x:Name="VerboseOutput" Margin="10,0,0,0" IsChecked="True" />
+            <CheckBox Grid.Row="1" Grid.Column="2" x:Name="NoFileSizes" Margin="10,0,0,0" />
+            <CheckBox Grid.Row="2" Grid.Column="2" x:Name="NoProgress" Margin="10,0,0,0" />
+            <CheckBox Grid.Row="3" Grid.Column="2" x:Name="ChkListOnly" Margin="10,0,0,0" />
+
+        </Grid>
+</Expander>
