+++ RoboSharp.BackupApp/UserControls/MultiJob_CommandProgressIndicator.xaml.cs
